CREATE BACKEND & FRONTEND MODULES
Creates both backend & frontend modules. Moves front's feature from `core` to `frontend`.
