--- core/build.gradle
@@ -1,6 +1,0 @@
-dependencies {
-    testCompile "com.badlogicgames.gdx:gdx-backend-headless:$gdxVersion"
-    testCompile "com.badlogicgames.gdx:gdx-platform:$gdxVersion:natives-desktop"
-    compile "com.badlogicgames.gdx:gdx:$gdxVersion"
-}
-
