+++ core/src/main/resources/loader_test.properties
