--- core/src/test/java/fr/foacs/ribz/core/mocks/AssetManagerMock.java
@@ -1,61 +1,0 @@
-/*
- * Copyright or © or Copr. Foacs
- * contributor(s):
- *  - Alexis DINQUER adinquer@yahoo.com
- *  - Brice KESSLER brice.kessler@outlook.com
- *
- * This software is a computer program whose purpose is to develop and
- * play board game.
- *
- * This software is governed by the CeCILL license under French law and
- * abiding by the rules of distribution of free software. You can use,
- * modify and/ or redistribute the software under the terms of the CeCILL
- * license as circulated by CEA, CNRS and INRIA at the following URL
- * "http://www.cecill.info".
- *
- * As a counterpart to the access to the source code and rights to copy,
- * modify and redistribute granted by the license, users are provided only
- * with a limited warranty and the software's author, the holder of the
- * economic rights, and the successive licensors have only limited
- * liability.
- *
- * In this respect, the user's attention is drawn to the risks associated
- * with loading, using, modifying and/or developing or reproducing the
- * software by the user in light of its specific status of free software,
- * that may mean that it is complicated to manipulate, and that also
- * therefore means that it is reserved for developers and experienced
- * professionals having in-depth computer knowledge. Users are therefore
- * encourage to load and test the software's suitability as regards their
- * requirements in conditions enabling the security of their systems and/or
- * data to be ensured and, more generally, to use and operate it in the
- * same conditions as regards security.
- *
- * The fact that you are presently reading this means that you have had
- * knowledge of the CeCILL license and that you accept its terms.
- */
-
-package fr.foacs.ribz.core.mocks;
-
-import com.badlogic.gdx.assets.AssetManager;
-import com.badlogic.gdx.maps.tiled.TiledMap;
-
-/**
- * Mock class for the {@link AssetManager} class.
- */
-public class AssetManagerMock extends AssetManager {
-
-  TiledMap map = new TiledMapMock();
-
-
-  @Override
-  public synchronized <T> T get(String fileName, Class<T> type) {
-    if (type.getTypeName().equals(TiledMap.class.getTypeName())) {
-      return (T) map;
-    }
-    return super.get(fileName, type);
-  }
-
-  public TiledMap getMockedTiledMap() {
-    return map;
-  }
-}
