+++ desktop/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
-    compile project(":core")
+    compile project(":backend")
+    compile project(":frontend")
     compile "com.badlogicgames.gdx:gdx-backend-lwjgl:$gdxVersion"
     compile "com.badlogicgames.gdx:gdx-platform:$gdxVersion:natives-desktop"
     compile "com.badlogicgames.gdx:gdx-box2d-platform:$gdxVersion:natives-desktop"
