+++ frontend/src/main/resources/colors.properties
