+++ frontend/src/test/java/fr/foacs/ribz/frontend/mocks/OrthographicCameraMock.java
@@ -1,8 +1,8 @@
 /*
  * Copyright or © or Copr. Foacs
  * contributor(s):
- * - Alexis DINQUER adinquer@yahoo.com
- * - Brice KESSLER brice.kessler@outlook.com
+ *  - Alexis DINQUER adinquer@yahoo.com
+ *  - Brice KESSLER brice.kessler@outlook.com
  *
  * This software is a computer program whose purpose is to develop and
  * play board game.
@@ -22,7 +22,7 @@
  * In this respect, the user's attention is drawn to the risks associated
  * with loading, using, modifying and/or developing or reproducing the
  * software by the user in light of its specific status of free software,
- * that may mean that it is complicated to manipulaten, and that also
+ * that may mean that it is complicated to manipulate, and that also
  * therefore means that it is reserved for developers and experienced
  * professionals having in-depth computer knowledge. Users are therefore
  * encourage to load and test the software's suitability as regards their
@@ -34,7 +34,18 @@
  * knowledge of the CeCILL license and that you accept its terms.
  */
 
-/**
- * Contains tests for desktop module.
- */
-package fr.foacs.ribz.desktop;
+package fr.foacs.ribz.frontend.mocks;
+
+import com.badlogic.gdx.graphics.OrthographicCamera;
+
+public class OrthographicCameraMock extends OrthographicCamera {
+
+  public OrthographicCameraMock() {
+    // Avoid calling parent constructor
+  }
+
+  @Override
+  public void update() {
+    // Avoid calling parent update();
+  }
+}
