+++ frontend/src/test/java/fr/foacs/ribz/frontend/mocks/TiledMapMock.java
@@ -1,8 +1,8 @@
 /*
  * Copyright or © or Copr. Foacs
  * contributor(s):
- * - Alexis DINQUER adinquer@yahoo.com
- * - Brice KESSLER brice.kessler@outlook.com
+ *  - Alexis DINQUER adinquer@yahoo.com
+ *  - Brice KESSLER brice.kessler@outlook.com
  *
  * This software is a computer program whose purpose is to develop and
  * play board game.
@@ -22,7 +22,7 @@
  * In this respect, the user's attention is drawn to the risks associated
  * with loading, using, modifying and/or developing or reproducing the
  * software by the user in light of its specific status of free software,
- * that may mean that it is complicated to manipulaten, and that also
+ * that may mean that it is complicated to manipulate, and that also
  * therefore means that it is reserved for developers and experienced
  * professionals having in-depth computer knowledge. Users are therefore
  * encourage to load and test the software's suitability as regards their
@@ -34,7 +34,31 @@
  * knowledge of the CeCILL license and that you accept its terms.
  */
 
+package fr.foacs.ribz.frontend.mocks;
+
+import com.badlogic.gdx.maps.MapProperties;
+import com.badlogic.gdx.maps.tiled.TiledMap;
+
 /**
- * Contains tests for desktop module.
+ * Mock class for the {@link TiledMap} class
  */
-package fr.foacs.ribz.desktop;
+public class TiledMapMock extends TiledMap {
+
+  public static final int MAP_WIDTH = 15;
+  public static final int MAP_HEIGHT = 22;
+  public static final int TILE_WIDTH = 32;
+  public static final int TILE_HEIGHT = 36;
+  private static final MapProperties PROPERTIES = new MapProperties();
+
+  static {
+    PROPERTIES.put("width", MAP_WIDTH);
+    PROPERTIES.put("height", MAP_HEIGHT);
+    PROPERTIES.put("tilewidth", TILE_WIDTH);
+    PROPERTIES.put("tileheight", TILE_HEIGHT);
+  }
+
+  @Override
+  public MapProperties getProperties() {
+    return PROPERTIES;
+  }
+}
