+++ frontend/src/test/java/fr/foacs/ribz/frontend/screens/BoardGameScreenTest.java
@@ -22,7 +22,7 @@
  * In this respect, the user's attention is drawn to the risks associated
  * with loading, using, modifying and/or developing or reproducing the
  * software by the user in light of its specific status of free software,
- * that may mean that it is complicated to manipulaten, and that also
+ * that may mean that it is complicated to manipulate, and that also
  * therefore means that it is reserved for developers and experienced
  * professionals having in-depth computer knowledge. Users are therefore
  * encourage to load and test the software's suitability as regards their
@@ -34,7 +34,33 @@
  * knowledge of the CeCILL license and that you accept its terms.
  */
 
+package fr.foacs.ribz.frontend.screens;
+
+import com.badlogic.gdx.Screen;
+import fr.foacs.ribz.frontend.controllers.BoardGameController;
+import fr.foacs.ribz.frontend.mocks.BoardControllerMock;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 /**
- * Contains tests for desktop module.
+ * Test class  for the {@link BoardGameScreens} class.
  */
-package fr.foacs.ribz.desktop;
+@DisplayName("BoardGameScreen")
+class BoardGameScreenTest {
+
+  /**
+   * Test the {@link BoardGameScreens#createScreen(BoardGameController)} method with board.
+   */
+  @DisplayName("Create screen: with board")
+  @Test
+  void testCreateScreenWithBoard() {
+    final BoardGameController controller = new BoardControllerMock();
+    final Screen screen = BoardGameScreens.BOARD.createScreen(controller);
+    assertNotNull(screen);
+    assertTrue(screen instanceof BoardScreen);
+  }
+
+}
